Vertex is now its own class
It is mutable and knows about its position in different spaces. Not
sure if this is ultimately the best way to implement this but it'll do
for now. Anticipate additional data being added to the
ObjectWithVertices class (which needs a rename) such as normal
vectors and shapes which reference the vertices.