Bring back the weird totem pole thingie
This results in background objects being drawn over foreground objects,
so we sort the objects in descending order of z coord (in view space).
Sadly this doesn't work brilliantly for compound objects (which may
contain multiple particles) so next it's going to be necessary to sort all
of the individual particles (and eventually polygons etc).