Uses an affine transform matrix to map between screen and argand plane
Previously used hand-cranked dodgy maths that could only scale and
translate. This way is more general (and will allow us to implemen
rotations). It might be slightly slower, though, because it has to do
six multiplications per pixel instead of two. But those multiplications
shouldn't be on the critical path (we do waaaay more multiplications
for each pixel just to try and work out if that pixel is in the Mandelbrot
set or not).